Extract module Code
